wasm: fix u32 overflow in BufferedFileReader for offsets above 4GB
* In WASM usize is 32-bit, so casting u64 offsets to usize truncated
  values beyond 4GB, causing "attempt to add with overflow" panic
  when reading multi-file packages (pck + pkx) larger than 4GB

+++ autoangel-wasm/src/file_reader.rs
@@ -44,8 +44,8 @@ impl BufferedFileReader {
     }
 
     async fn fetch_chunk(&self, offset: u64, min_len: usize) -> Result<()> {
-        let chunk_size = self.chunk_size.max(min_len);
-        let end = ((offset as usize) + chunk_size).min(self.size as usize);
+        let chunk_size = self.chunk_size.max(min_len) as u64;
+        let end = (offset + chunk_size).min(self.size);
         let start = offset as f64;
 
         // File inherits from Blob, use Blob::slice
